+++ src/mako/syringe/exceptions/ContainerException.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\syringe\exceptions;
+
+use RuntimeException;
+
+/**
+ * Container exception.
+ *
+ * @author Frederic G. Østby
+ */
+class ContainerException extends RuntimeException
+{
+
+}