+++ src/mako/syringe/exceptions/UnableToInstantiateException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\syringe\exceptions;
+
+/**
+ * Unable to instantiate exception.
+ *
+ * @author Frederic G. Østby
+ */
+class UnableToInstantiateException extends ContainerException
+{
+
+}