+++ src/mako/syringe/exceptions/UnableToResolveParameterException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\syringe\exceptions;
+
+/**
+ * Unable to resolve parameter exception.
+ *
+ * @author Frederic G. Østby
+ */
+class UnableToResolveParameterException extends ContainerException
+{
+
+}