+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The major version bump is due to dropped support for PHP `7.0` and `7.1`. Applic
 * Removed the deprecated `FileSystem::hash()` method.
 * Removed the deprecated `FileSystem::hmac()` method.
 * The `RequestException` class has been renamed to `HttpException`.
+* The container now supports parameters that are nullable but not optional.
+* The container will no longer throw an exception when unable to resolve a nullable or optional class dependency.
 
 > Check out the upgrade guide for details on how to upgrade from `5.7.*.`